dataops_assistent_backend/requirements.txt
fastapi
uvicorn
openai
python-dotenv
jsonschema
pandas
minio
boto3
python-multipart